Refactor app and reducer internals
