+++ src/peneo/app.py
@@ -2,26 +2,27 @@
 
 import threading
 from collections.abc import Sequence
-from concurrent.futures import CancelledError as FutureCancelledError
-from contextlib import nullcontext
-from functools import partial
 from pathlib import Path
 
 from textual import events
-from textual.app import App, ComposeResult, SuspendNotSupported
+from textual.app import App, ComposeResult
 from textual.binding import Binding
-from textual.containers import Horizontal, Vertical
+from textual.containers import Vertical
 from textual.css.query import NoMatches
 from textual.message import Message
 from textual.timer import Timer
-from textual.worker import Worker, WorkerState
+from textual.worker import Worker
 
 from peneo.adapters import LocalExternalLaunchAdapter
+from peneo.app_runtime import (
+    cancel_pending_runtime_work,
+    handle_worker_state_changed,
+    schedule_effects,
+    sync_runtime_state,
+)
+from peneo.app_shell import build_body, refresh_shell, resize_split_terminal_session
 from peneo.models import (
     AppConfig,
-    FileMutationResult,
-    PasteConflictPrompt,
-    PasteExecutionResult,
     ThreePaneShellData,
 )
 from peneo.services import (
@@ -33,8 +34,6 @@ from peneo.services import (
     FileMutationService,
     FileSearchService,
     GrepSearchService,
-    InvalidFileSearchQueryError,
-    InvalidGrepSearchQueryError,
     LiveBrowserSnapshotLoader,
     LiveClipboardOperationService,
     LiveConfigSaveService,
@@ -51,49 +50,14 @@ from peneo.services import (
 from peneo.state import (
     Action,
     AppState,
-    BrowserSnapshotFailed,
-    BrowserSnapshotLoaded,
-    ChildPaneSnapshotFailed,
-    ChildPaneSnapshotLoaded,
-    ClipboardPasteCompleted,
-    ClipboardPasteFailed,
-    ClipboardPasteNeedsResolution,
-    CloseSplitTerminalEffect,
-    ConfigSaveCompleted,
-    ConfigSaveFailed,
-    DirectorySizesFailed,
-    DirectorySizesLoaded,
     Effect,
     ExitCurrentPath,
-    ExternalLaunchCompleted,
-    ExternalLaunchFailed,
-    FileMutationCompleted,
-    FileMutationFailed,
-    FileSearchCompleted,
-    FileSearchFailed,
-    GrepSearchCompleted,
-    GrepSearchFailed,
-    LoadBrowserSnapshotEffect,
-    LoadChildPaneSnapshotEffect,
     NotificationState,
-    PasteFromClipboardEffect,
     ReduceResult,
     RequestBrowserSnapshot,
-    RunClipboardPasteEffect,
-    RunConfigSaveEffect,
-    RunDirectorySizeEffect,
-    RunExternalLaunchEffect,
-    RunFileMutationEffect,
-    RunFileSearchEffect,
-    RunGrepSearchEffect,
-    SetNotification,
     SetTerminalHeight,
     SortState,
     SplitTerminalExited,
-    SplitTerminalStarted,
-    SplitTerminalStartFailed,
-    StartSplitTerminalEffect,
-    WriteSplitTerminalInputEffect,
     build_placeholder_app_state,
     dispatch_key_input,
     iter_bound_keys,
@@ -107,14 +71,9 @@ from peneo.ui import (
     ConflictDialog,
     CurrentPathBar,
     HelpBar,
-    MainPane,
-    SidePane,
     SplitTerminalPane,
     StatusBar,
 )
-
-_FILE_SEARCH_DEBOUNCE_SECONDS = 0.2
-_GREP_SEARCH_DEBOUNCE_SECONDS = 0.2
 
 
 class PeneoApp(App[None]):
@@ -435,9 +394,7 @@ class PeneoApp(App[None]):
     def on_unmount(self) -> None:
         """Ensure the embedded terminal session is stopped when the app exits."""
 
-        self._cancel_pending_file_search()
-        self._cancel_pending_grep_search()
-        self._cancel_pending_directory_size()
+        cancel_pending_runtime_work(self)
         if self._split_terminal_session is None:
             return
         self._split_terminal_session.close()
@@ -471,34 +428,7 @@ class PeneoApp(App[None]):
         await self._dispatch_key_press(key, character=character)
 
     def _build_body(self, shell: ThreePaneShellData) -> Vertical:
-        return Vertical(
-            Horizontal(
-                SidePane(
-                    "Parent Directory",
-                    shell.parent_entries,
-                    id="parent-pane",
-                    classes="pane side-pane",
-                ),
-                MainPane(
-                    "Current Directory",
-                    shell.current_entries,
-                    summary=shell.current_summary,
-                    cursor_index=shell.current_cursor_index,
-                    context_input=shell.current_context_input,
-                    id="current-pane",
-                    classes="pane main-pane",
-                ),
-                SidePane(
-                    "Child Directory",
-                    shell.child_entries,
-                    id="child-pane",
-                    classes="pane side-pane",
-                ),
-                id="browser-row",
-            ),
-            SplitTerminalPane(shell.split_terminal, id="split-terminal"),
-            id="body",
-        )
+        return build_body(shell)
 
     async def _dispatch_key_press(
         self,
@@ -521,16 +451,14 @@ class PeneoApp(App[None]):
 
         previous_state = self._app_state
         changed, effects = self._apply_actions(actions)
-        self._sync_file_search_state(previous_state, self._app_state)
-        self._sync_grep_search_state(previous_state, self._app_state)
-        self._sync_directory_size_state(previous_state, self._app_state)
+        sync_runtime_state(self, previous_state, self._app_state)
         if previous_state.config.display.theme != self._app_state.config.display.theme:
             self.theme = self._app_state.config.display.theme
         if previous_state.config != self._app_state.config:
             self._sync_external_launch_service()
         if changed:
             await self._refresh_shell()
-        self._schedule_effects(effects)
+        schedule_effects(self, effects)
 
     def _build_external_launch_service(self, app_config: AppConfig) -> LiveExternalLaunchService:
         return LiveExternalLaunchService(
@@ -564,429 +492,6 @@ class PeneoApp(App[None]):
         self._app_state = state
         return changed, tuple(effects)
 
-    def _schedule_effects(self, effects: Sequence[Effect]) -> None:
-        for effect in effects:
-            if isinstance(effect, LoadBrowserSnapshotEffect):
-                self._schedule_browser_snapshot(effect)
-            elif isinstance(effect, LoadChildPaneSnapshotEffect):
-                self._schedule_child_pane_snapshot(effect)
-            elif isinstance(effect, RunClipboardPasteEffect):
-                self._schedule_clipboard_paste(effect)
-            elif isinstance(effect, RunConfigSaveEffect):
-                self._schedule_config_save(effect)
-            elif isinstance(effect, RunDirectorySizeEffect):
-                self._schedule_directory_sizes(effect)
-            elif isinstance(effect, RunFileMutationEffect):
-                self._schedule_file_mutation(effect)
-            elif isinstance(effect, RunExternalLaunchEffect):
-                if effect.request.kind == "copy_paths":
-                    self._run_copy_paths(effect)
-                elif effect.request.kind == "open_editor":
-                    self.call_next(self._run_foreground_external_launch, effect)
-                else:
-                    self._schedule_external_launch(effect)
-            elif isinstance(effect, RunFileSearchEffect):
-                self._schedule_file_search(effect)
-            elif isinstance(effect, RunGrepSearchEffect):
-                self._schedule_grep_search(effect)
-            elif isinstance(effect, StartSplitTerminalEffect):
-                self._start_split_terminal(effect)
-            elif isinstance(effect, WriteSplitTerminalInputEffect):
-                self._write_split_terminal_input(effect)
-            elif isinstance(effect, PasteFromClipboardEffect):
-                self._paste_from_clipboard(effect)
-            elif isinstance(effect, CloseSplitTerminalEffect):
-                self._close_split_terminal(effect)
-
-    def _schedule_browser_snapshot(self, effect: LoadBrowserSnapshotEffect) -> None:
-        worker = self.run_worker(
-            partial(
-                self._snapshot_loader.load_browser_snapshot,
-                effect.path,
-                effect.cursor_path,
-            ),
-            name=f"browser-snapshot:{effect.request_id}",
-            group="browser-snapshot",
-            description=effect.path,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_child_pane_snapshot(self, effect: LoadChildPaneSnapshotEffect) -> None:
-        worker = self.run_worker(
-            partial(
-                self._snapshot_loader.load_child_pane_snapshot,
-                effect.current_path,
-                effect.cursor_path,
-            ),
-            name=f"child-pane-snapshot:{effect.request_id}",
-            group="child-pane-snapshot",
-            description=effect.cursor_path,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_clipboard_paste(self, effect: RunClipboardPasteEffect) -> None:
-        worker = self.run_worker(
-            partial(self._clipboard_service.execute_paste, effect.request),
-            name=f"clipboard-paste:{effect.request_id}",
-            group="clipboard-paste",
-            description=effect.request.destination_dir,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_config_save(self, effect: RunConfigSaveEffect) -> None:
-        worker = self.run_worker(
-            partial(
-                self._config_save_service.save,
-                path=effect.path,
-                config=effect.config,
-            ),
-            name=f"config-save:{effect.request_id}",
-            group="config-save",
-            description=effect.path,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_directory_sizes(self, effect: RunDirectorySizeEffect) -> None:
-        cancel_event = threading.Event()
-        self._active_directory_size_cancel_event = cancel_event
-        self._active_directory_size_request_id = effect.request_id
-        worker = self.run_worker(
-            partial(
-                self._directory_size_service.calculate_sizes,
-                effect.paths,
-                is_cancelled=cancel_event.is_set,
-            ),
-            name=f"directory-size:{effect.request_id}",
-            group="directory-size",
-            description=",".join(effect.paths),
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_file_mutation(self, effect: RunFileMutationEffect) -> None:
-        worker = self.run_worker(
-            partial(self._file_mutation_service.execute, effect.request),
-            name=f"file-mutation:{effect.request_id}",
-            group="file-mutation",
-            description=str(effect.request),
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_external_launch(self, effect: RunExternalLaunchEffect) -> None:
-        worker = self.run_worker(
-            partial(self._external_launch_service.execute, effect.request),
-            name=f"external-launch:{effect.request_id}",
-            group="external-launch",
-            description=str(effect.request),
-            exit_on_error=False,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_file_search(self, effect: RunFileSearchEffect) -> None:
-        self._cancel_file_search_timer()
-        self._file_search_timer = self.set_timer(
-            _FILE_SEARCH_DEBOUNCE_SECONDS,
-            partial(self._start_file_search_worker, effect),
-            name=f"file-search-debounce:{effect.request_id}",
-        )
-
-    def _start_file_search_worker(self, effect: RunFileSearchEffect) -> None:
-        self._file_search_timer = None
-        if self._app_state.pending_file_search_request_id != effect.request_id:
-            return
-        cancel_event = threading.Event()
-        self._active_file_search_cancel_event = cancel_event
-        self._active_file_search_request_id = effect.request_id
-        worker = self.run_worker(
-            partial(
-                self._file_search_service.search,
-                effect.root_path,
-                effect.query,
-                show_hidden=effect.show_hidden,
-                is_cancelled=cancel_event.is_set,
-            ),
-            name=f"file-search:{effect.request_id}",
-            group="file-search",
-            description=effect.query,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _schedule_grep_search(self, effect: RunGrepSearchEffect) -> None:
-        self._cancel_grep_search_timer()
-        self._grep_search_timer = self.set_timer(
-            _GREP_SEARCH_DEBOUNCE_SECONDS,
-            partial(self._start_grep_search_worker, effect),
-            name=f"grep-search-debounce:{effect.request_id}",
-        )
-
-    def _start_grep_search_worker(self, effect: RunGrepSearchEffect) -> None:
-        self._grep_search_timer = None
-        if self._app_state.pending_grep_search_request_id != effect.request_id:
-            return
-        cancel_event = threading.Event()
-        self._active_grep_search_cancel_event = cancel_event
-        self._active_grep_search_request_id = effect.request_id
-        worker = self.run_worker(
-            partial(
-                self._grep_search_service.search,
-                effect.root_path,
-                effect.query,
-                show_hidden=effect.show_hidden,
-                is_cancelled=cancel_event.is_set,
-            ),
-            name=f"grep-search:{effect.request_id}",
-            group="grep-search",
-            description=effect.query,
-            exit_on_error=False,
-            exclusive=True,
-            thread=True,
-        )
-        self._pending_workers[worker.name] = effect
-
-    def _sync_file_search_state(self, previous_state: AppState, next_state: AppState) -> None:
-        if (
-            previous_state.pending_file_search_request_id
-            == next_state.pending_file_search_request_id
-        ):
-            return
-        self._cancel_pending_file_search()
-
-    def _sync_grep_search_state(self, previous_state: AppState, next_state: AppState) -> None:
-        if (
-            previous_state.pending_grep_search_request_id
-            == next_state.pending_grep_search_request_id
-        ):
-            return
-        self._cancel_pending_grep_search()
-
-    def _sync_directory_size_state(self, previous_state: AppState, next_state: AppState) -> None:
-        if (
-            previous_state.pending_directory_size_request_id
-            == next_state.pending_directory_size_request_id
-        ):
-            return
-        self._cancel_pending_directory_size()
-
-    def _cancel_pending_file_search(self) -> None:
-        self._cancel_file_search_timer()
-        self._cancel_active_file_search()
-
-    def _cancel_file_search_timer(self) -> None:
-        if self._file_search_timer is None:
-            return
-        self._file_search_timer.stop()
-        self._file_search_timer = None
-
-    def _cancel_active_file_search(self) -> None:
-        if self._active_file_search_cancel_event is None:
-            return
-        self._active_file_search_cancel_event.set()
-        self._active_file_search_cancel_event = None
-        self._active_file_search_request_id = None
-
-    def _cancel_pending_grep_search(self) -> None:
-        self._cancel_grep_search_timer()
-        self._cancel_active_grep_search()
-
-    def _cancel_grep_search_timer(self) -> None:
-        if self._grep_search_timer is None:
-            return
-        self._grep_search_timer.stop()
-        self._grep_search_timer = None
-
-    def _cancel_active_grep_search(self) -> None:
-        if self._active_grep_search_cancel_event is None:
-            return
-        self._active_grep_search_cancel_event.set()
-        self._active_grep_search_cancel_event = None
-        self._active_grep_search_request_id = None
-
-    def _cancel_pending_directory_size(self) -> None:
-        if self._active_directory_size_cancel_event is None:
-            return
-        self._active_directory_size_cancel_event.set()
-        self._active_directory_size_cancel_event = None
-        self._active_directory_size_request_id = None
-
-    def _start_split_terminal(self, effect: StartSplitTerminalEffect) -> None:
-        try:
-            session = self._split_terminal_service.start(
-                effect.cwd,
-                on_output=partial(self._handle_split_terminal_output, effect.session_id),
-                on_exit=partial(self._handle_split_terminal_exit, effect.session_id),
-            )
-        except OSError as error:
-            self.call_next(
-                self.dispatch_actions,
-                (
-                    SplitTerminalStartFailed(
-                        session_id=effect.session_id,
-                        message=str(error) or "Failed to open split terminal",
-                    ),
-                ),
-            )
-            return
-
-        self._split_terminal_session = session
-        self.call_next(
-            self.dispatch_actions,
-            (
-                SplitTerminalStarted(session_id=effect.session_id, cwd=effect.cwd),
-            ),
-        )
-
-    def _write_split_terminal_input(self, effect: WriteSplitTerminalInputEffect) -> None:
-        if self._app_state.split_terminal.session_id != effect.session_id:
-            return
-        if self._split_terminal_session is None:
-            return
-        try:
-            self._split_terminal_session.write(effect.data)
-        except OSError as error:
-            self.call_next(
-                self.dispatch_actions,
-                (
-                    SplitTerminalStartFailed(
-                        session_id=effect.session_id,
-                        message=str(error) or "Failed to write to split terminal",
-                    ),
-                ),
-            )
-
-    def _paste_from_clipboard(self, effect: PasteFromClipboardEffect) -> None:
-        if self._app_state.split_terminal.session_id != effect.session_id:
-            return
-        if self._split_terminal_session is None:
-            return
-        try:
-            clipboard_text = self._external_launch_service.get_from_clipboard()
-            self._split_terminal_session.write(clipboard_text)
-        except OSError as error:
-            message = str(error) or "Failed to read clipboard"
-            self.dispatch_actions(
-                SetNotification(NotificationState(level="warning", message=message)),
-            )
-
-    def _close_split_terminal(self, effect: CloseSplitTerminalEffect) -> None:
-        if self._split_terminal_session is None:
-            return
-        try:
-            self._split_terminal_session.close()
-        finally:
-            self._split_terminal_session = None
-
-    def _run_foreground_external_launch(self, effect: RunExternalLaunchEffect) -> None:
-        suspend_context = nullcontext()
-        try:
-            suspend_context = self.suspend()
-        except SuspendNotSupported as error:
-            self.call_next(
-                self.dispatch_actions,
-                (
-                    ExternalLaunchFailed(
-                        request_id=effect.request_id,
-                        request=effect.request,
-                        message=str(error),
-                    ),
-                ),
-            )
-            return
-
-        try:
-            with suspend_context:
-                self._external_launch_service.execute(effect.request)
-        except OSError as error:
-            self.refresh(repaint=True, layout=True)
-            self.call_next(
-                self.dispatch_actions,
-                (
-                    ExternalLaunchFailed(
-                        request_id=effect.request_id,
-                        request=effect.request,
-                        message=str(error) or "Operation failed",
-                    ),
-                ),
-            )
-            return
-
-        self.refresh(repaint=True, layout=True)
-        self.call_next(
-            self.dispatch_actions,
-            (
-                ExternalLaunchCompleted(
-                    request_id=effect.request_id,
-                    request=effect.request,
-                ),
-            ),
-        )
-
-    def _run_copy_paths(self, effect: RunExternalLaunchEffect) -> None:
-        try:
-            self._external_launch_service.execute(effect.request)
-        except OSError as error:
-            self.call_next(
-                self.dispatch_actions,
-                (
-                    ExternalLaunchFailed(
-                        request_id=effect.request_id,
-                        request=effect.request,
-                        message=str(error) or "Operation failed",
-                    ),
-                ),
-            )
-            return
-
-        self.call_next(
-            self.dispatch_actions,
-            (
-                ExternalLaunchCompleted(
-                    request_id=effect.request_id,
-                    request=effect.request,
-                ),
-            ),
-        )
-
-    def _handle_split_terminal_output(self, session_id: int, data: str) -> None:
-        message = self.SplitTerminalOutput(session_id=session_id, data=data)
-        try:
-            if self._thread_id == threading.get_ident():
-                self.post_message(message)
-                return
-            self.call_from_thread(self.post_message, message)
-        except (RuntimeError, FutureCancelledError):
-            return
-
-    def _handle_split_terminal_exit(self, session_id: int, exit_code: int | None) -> None:
-        message = self.SplitTerminalExitedMessage(session_id=session_id, exit_code=exit_code)
-        try:
-            if self._thread_id == threading.get_ident():
-                self.post_message(message)
-                return
-            self.call_from_thread(self.post_message, message)
-        except (RuntimeError, FutureCancelledError):
-            return
-
     async def on_peneo_app_split_terminal_output(
         self,
         message: SplitTerminalOutput,
@@ -1018,262 +523,7 @@ class PeneoApp(App[None]):
         )
 
     async def on_worker_state_changed(self, event: Worker.StateChanged) -> None:
-        """Convert worker completion back into reducer actions."""
-
-        effect = self._pending_workers.get(event.worker.name)
-        if effect is None:
-            return
-
-        if event.state in {WorkerState.PENDING, WorkerState.RUNNING}:
-            return
-
-        self._pending_workers.pop(event.worker.name, None)
-        if (
-            isinstance(effect, RunFileSearchEffect)
-            and effect.request_id == self._active_file_search_request_id
-        ):
-            self._active_file_search_cancel_event = None
-            self._active_file_search_request_id = None
-        if (
-            isinstance(effect, RunGrepSearchEffect)
-            and effect.request_id == self._active_grep_search_request_id
-        ):
-            self._active_grep_search_cancel_event = None
-            self._active_grep_search_request_id = None
-        if (
-            isinstance(effect, RunDirectorySizeEffect)
-            and effect.request_id == self._active_directory_size_request_id
-        ):
-            self._active_directory_size_cancel_event = None
-            self._active_directory_size_request_id = None
-
-        if event.state == WorkerState.CANCELLED:
-            return
-
-        if event.state == WorkerState.SUCCESS:
-            if isinstance(effect, LoadBrowserSnapshotEffect):
-                await self.dispatch_actions(
-                    (
-                        BrowserSnapshotLoaded(
-                            request_id=effect.request_id,
-                            snapshot=event.worker.result,
-                            blocking=effect.blocking,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, LoadChildPaneSnapshotEffect):
-                await self.dispatch_actions(
-                    (
-                        ChildPaneSnapshotLoaded(
-                            request_id=effect.request_id,
-                            pane=event.worker.result,
-                        ),
-                    )
-                )
-                return
-
-            result = event.worker.result
-            if isinstance(result, PasteConflictPrompt):
-                await self.dispatch_actions(
-                    (
-                        ClipboardPasteNeedsResolution(
-                            request_id=effect.request_id,
-                            request=result.request,
-                            conflicts=result.conflicts,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(result, PasteExecutionResult):
-                await self.dispatch_actions(
-                    (
-                        ClipboardPasteCompleted(
-                            request_id=effect.request_id,
-                            summary=result.summary,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(result, FileMutationResult):
-                await self.dispatch_actions(
-                    (
-                        FileMutationCompleted(
-                            request_id=effect.request_id,
-                            result=result,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, RunConfigSaveEffect):
-                await self.dispatch_actions(
-                    (
-                        ConfigSaveCompleted(
-                            request_id=effect.request_id,
-                            path=event.worker.result,
-                            config=effect.config,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, RunDirectorySizeEffect):
-                await self.dispatch_actions(
-                    (
-                        DirectorySizesLoaded(
-                            request_id=effect.request_id,
-                            sizes=event.worker.result[0],
-                            failures=event.worker.result[1],
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, RunExternalLaunchEffect):
-                await self.dispatch_actions(
-                    (
-                        ExternalLaunchCompleted(
-                            request_id=effect.request_id,
-                            request=effect.request,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, RunFileSearchEffect):
-                await self.dispatch_actions(
-                    (
-                        FileSearchCompleted(
-                            request_id=effect.request_id,
-                            query=effect.query,
-                            results=event.worker.result,
-                        ),
-                    )
-                )
-                return
-
-            if isinstance(effect, RunGrepSearchEffect):
-                await self.dispatch_actions(
-                    (
-                        GrepSearchCompleted(
-                            request_id=effect.request_id,
-                            query=effect.query,
-                            results=event.worker.result,
-                        ),
-                    )
-                )
-                return
-
-        message = str(event.worker.error) or "Operation failed"
-        if isinstance(effect, LoadBrowserSnapshotEffect):
-            await self.dispatch_actions(
-                (
-                    BrowserSnapshotFailed(
-                        request_id=effect.request_id,
-                        message=message,
-                        blocking=effect.blocking,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, LoadChildPaneSnapshotEffect):
-            await self.dispatch_actions(
-                (
-                    ChildPaneSnapshotFailed(
-                        request_id=effect.request_id,
-                        message=message,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunFileMutationEffect):
-            await self.dispatch_actions(
-                (
-                    FileMutationFailed(
-                        request_id=effect.request_id,
-                        message=message,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunConfigSaveEffect):
-            await self.dispatch_actions(
-                (
-                    ConfigSaveFailed(
-                        request_id=effect.request_id,
-                        message=message,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunDirectorySizeEffect):
-            await self.dispatch_actions(
-                (
-                    DirectorySizesFailed(
-                        request_id=effect.request_id,
-                        paths=effect.paths,
-                        message=message,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunExternalLaunchEffect):
-            await self.dispatch_actions(
-                (
-                    ExternalLaunchFailed(
-                        request_id=effect.request_id,
-                        request=effect.request,
-                        message=message,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunFileSearchEffect):
-            invalid_query = isinstance(event.worker.error, InvalidFileSearchQueryError)
-            await self.dispatch_actions(
-                (
-                    FileSearchFailed(
-                        request_id=effect.request_id,
-                        query=effect.query,
-                        message=message,
-                        invalid_query=invalid_query,
-                    ),
-                )
-            )
-            return
-
-        if isinstance(effect, RunGrepSearchEffect):
-            invalid_query = isinstance(event.worker.error, InvalidGrepSearchQueryError)
-            await self.dispatch_actions(
-                (
-                    GrepSearchFailed(
-                        request_id=effect.request_id,
-                        query=effect.query,
-                        message=message,
-                        invalid_query=invalid_query,
-                    ),
-                )
-            )
-            return
-
-        await self.dispatch_actions(
-            (
-                ClipboardPasteFailed(
-                    request_id=effect.request_id,
-                    message=message,
-                ),
-            )
-        )
+        await handle_worker_state_changed(self, event)
 
     async def on_resize(self, event: events.Resize) -> None:
         """Keep the split-terminal PTY dimensions roughly aligned with the viewport."""
@@ -1285,82 +535,19 @@ class PeneoApp(App[None]):
         self.call_after_refresh(self._resize_split_terminal_session)
 
     async def _refresh_shell(self) -> None:
-        shell = select_shell_data(self._app_state)
-        try:
-            current_path_bar = self.query_one("#current-path-bar", CurrentPathBar)
-            parent_pane = self.query_one("#parent-pane", SidePane)
-            current_pane = self.query_one("#current-pane", MainPane)
-            child_pane = self.query_one("#child-pane", SidePane)
-            split_terminal = self.query_one("#split-terminal", SplitTerminalPane)
-            command_palette = self.query_one("#command-palette", CommandPalette)
-            help_bar = self.query_one("#help-bar", HelpBar)
-            status_bar = self.query_one("#status-bar", StatusBar)
-            conflict_dialog = self.query_one("#conflict-dialog", ConflictDialog)
-            attribute_dialog = self.query_one("#attribute-dialog", AttributeDialog)
-            config_dialog = self.query_one("#config-dialog", ConfigDialog)
-        except NoMatches:
-            selectors = (
-                "#current-path-bar",
-                "#body",
-                "#split-terminal",
-                "#command-palette",
-                "#help-bar",
-                "#status-bar",
-                "#conflict-dialog",
-                "#attribute-dialog",
-                "#config-dialog",
-            )
-            for selector in selectors:
-                try:
-                    await self.query_one(selector).remove()
-                except NoMatches:
-                    pass
-            await self.mount(CurrentPathBar(shell.current_path, id="current-path-bar"))
-            await self.mount(self._build_body(shell))
-            await self.mount(CommandPalette(shell.command_palette, id="command-palette"))
-            await self.mount(HelpBar(shell.help, id="help-bar"))
-            await self.mount(StatusBar(shell.status, id="status-bar"))
-            await self.mount(ConflictDialog(shell.conflict_dialog, id="conflict-dialog"))
-            await self.mount(AttributeDialog(shell.attribute_dialog, id="attribute-dialog"))
-            await self.mount(ConfigDialog(shell.config_dialog, id="config-dialog"))
-            return
-
-        current_path_bar.set_path(shell.current_path)
-        await parent_pane.set_entries(shell.parent_entries)
-        current_pane.set_summary(shell.current_summary)
-        current_pane.set_entries(shell.current_entries, shell.current_cursor_index)
-        current_pane.set_context_input(shell.current_context_input)
-        await child_pane.set_entries(shell.child_entries)
-        split_terminal.set_state(shell.split_terminal)
-        self._resize_split_terminal_session()
-        command_palette.set_state(shell.command_palette)
-        help_bar.set_state(shell.help)
-        status_bar.set_state(shell.status)
-        conflict_dialog.set_state(shell.conflict_dialog)
-        attribute_dialog.set_state(shell.attribute_dialog)
-        config_dialog.set_state(shell.config_dialog)
-
-        if (
-            self._app_state.ui_mode == "BROWSING"
-            and self._app_state.split_terminal.visible
-            and self._app_state.split_terminal.focus_target == "terminal"
-        ):
-            self.set_focus(split_terminal)
-        else:
-            try:
-                self.set_focus(current_pane.query_one("#current-pane-table"))
-            except NoMatches:
-                pass
+        await refresh_shell(
+            self,
+            self._app_state,
+            select_shell_data(self._app_state),
+            self._split_terminal_session,
+        )
 
     def _resize_split_terminal_session(self) -> None:
-        if self._split_terminal_session is None or not self._app_state.split_terminal.visible:
-            return
-        try:
-            split_terminal = self.query_one("#split-terminal", SplitTerminalPane)
-        except NoMatches:
-            return
-        columns, rows = split_terminal.terminal_dimensions()
-        self._split_terminal_session.resize(columns=columns, rows=rows)
+        resize_split_terminal_session(
+            self,
+            self._app_state,
+            self._split_terminal_session,
+        )
 
 
 def create_app(
